residents.js refactor
- refactored and modularized residents.js
